fix: fix simulation errors regarding APs
APPing and APStream used .doubleValue when .longValue should have been
used for parameter "stopAt". QoSReq could also experience overflow when
-1 (or "don't care") for maxAllowGap in AEs.

+++ src/DAF/AP/APPing/APPing.cc
@@ -101,7 +101,7 @@ void APPing::handleMessage(cMessage *msg) {
 
         }
         else if (!strcmp(msg->getName(), "ping")) {
-            if ((simTime().dbl()+1) < par("stopAt").doubleValue() ) {
+            if ((simTime().dbl()+1) < par("stopAt").longValue() ) {
                 connID = conIDsPing.front();
                 conIDsPing.pop();
